requirements.txt
fastapi
uvicorn
python-multipart
openai
langchain
langchain-community
langchain-openai
python-dotenv
